Move Logout
Moved logout component to auth directory and updated references.
