+++ src/app/app.module.ts
@@ -8,7 +8,7 @@ import { appRoutes } from './app.routing';
 import { AnonymousRequired, AuthRequired, AuthService, LoginButtonComponent } from './auth';
 import { environment } from '../environments/environment';
 import { HomeComponent } from './home';
-import { LogoutComponent } from './logout.component';
+import { LogoutComponent } from './auth/logout.component';
 import { NoContentComponent } from './no-content/no-content.component';
 import { Title } from '@angular/platform-browser';
 import { TodoService } from './todos/todos.service';
