+++ src/app/app.routing.ts
@@ -2,7 +2,7 @@ import { CanActivate, Routes } from '@angular/router';
 
 import { AnonymousRequired, AuthRequired } from './auth';
 import { HomeComponent } from './home/home.component';
-import { LogoutComponent } from './logout.component';
+import { LogoutComponent } from './auth/logout.component';
 import { NoContentComponent } from './no-content/no-content.component';
 
 export const appRoutes: Routes = [
